Move loadTrackable() from TrackableActivity to ConnectorFactory
Moved this function out of the TrackableActivity, as it'll be necessary
to use it again in LogTrackableActivity, for #5135 case (a).
